+++ frontend/v1/domain/entities/question/Suggestion.ts
@@ -6,55 +6,114 @@ import {
 } from "../IAnswer";
 import { QuestionType } from "./QuestionType";
 
+type AnswerValue = string | number | RankingAnswer | SpanAnswer;
+
+class SuggestionScore extends Number {
+  private constructor(value: number) {
+    super(value);
+  }
+
+  get fixed() {
+    return this.toFixed(1);
+  }
+
+  static from(value: number) {
+    return new SuggestionScore(value);
+  }
+}
+
+export class SuggestionValue {
+  public readonly score: SuggestionScore;
+  constructor(
+    public readonly value: AnswerValue,
+    score: number,
+    public readonly agent: string
+  ) {
+    this.score = score ? SuggestionScore.from(score) : undefined;
+  }
+}
+
 export class Suggestion implements Answer {
   constructor(
-    public readonly id: string,
+    private readonly id: string,
     public readonly questionId: string,
-    public readonly questionType: QuestionType,
-    public readonly suggestedAnswer: AnswerCombinations,
-    public readonly score: number,
-    public readonly agent: string
+    private readonly questionType: QuestionType,
+    private readonly suggestedAnswer: AnswerCombinations,
+    private readonly score: number | number[],
+    private readonly agent: string
   ) {}
 
   get value() {
     return this.suggestedAnswer;
   }
 
-  isSuggested(answer: string | number | RankingAnswer | SpanAnswer) {
+  isSuggested(answer: AnswerValue) {
     return !!this.getSuggestion(answer);
   }
 
-  getSuggestion(answer: string | number | RankingAnswer | SpanAnswer) {
+  getSuggestion(answer: AnswerValue): SuggestionValue | undefined {
     if (
       this.questionType.isSingleLabelType ||
       this.questionType.isTextType ||
       this.questionType.isRatingType
     ) {
-      return this.value === answer;
+      if (this.value === answer) {
+        return new SuggestionValue(answer, this.score as number, this.agent);
+      }
     }
 
     if (this.questionType.isMultiLabelType) {
       const multiLabel = this.value as string[];
       const answerValue = answer as string;
 
-      return multiLabel.includes(answerValue);
+      if (multiLabel.includes(answerValue)) {
+        const indexOf = multiLabel.indexOf(answerValue);
+
+        return new SuggestionValue(
+          answerValue,
+          this.score?.[indexOf],
+          this.agent
+        );
+      }
     }
 
     if (this.questionType.isSpanType) {
       const span = answer as SpanAnswer;
       const suggestions = this.value as SpanAnswer[];
 
-      return suggestions.find(
+      const spanSuggested = suggestions.find(
         (s) =>
           s.label === span.label && s.start === span.start && s.end === span.end
       );
+
+      if (spanSuggested) {
+        const indexOf = suggestions.indexOf(spanSuggested);
+
+        return new SuggestionValue(
+          spanSuggested,
+          this.score?.[indexOf],
+          this.agent
+        );
+      }
     }
 
     if (this.questionType.isRankingType) {
       const suggestedRanking = this.value as RankingAnswer[];
       const ranking = answer as RankingAnswer;
 
-      return suggestedRanking.find((s) => s.value === ranking.value);
+      const rankingSuggested = suggestedRanking.find(
+        (s) => s.value === ranking.value
+      );
+
+      if (rankingSuggested) {
+        const indexOf = suggestedRanking.indexOf(rankingSuggested);
+
+        return new SuggestionValue(
+          rankingSuggested,
+          this.score?.[indexOf],
+          this.agent
+        );
+      }
     }
   }
 }